delete

+++ Backend/BAMMMusic/Controllers/ArtistsController.cs
@@ -30,7 +30,7 @@ namespace BAMMMusic.Controllers
             return _context.Artists.ToList();
             
         }
-
+        
         // GET: api/Artists/5
         [HttpGet("{id}")]
         public async Task<IActionResult> GetArtist([FromRoute] int id)
@@ -98,12 +98,29 @@ namespace BAMMMusic.Controllers
         }
 
         // DELETE: api/Artists/5
-        [HttpDelete]
-        public ActionResult<IEnumerable<Artist>> Remove(Artist artist)
+        [HttpDelete("{id}")]
+        public async Task<IActionResult> DeleteArtist([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            var artist = await _context.Artists.FindAsync(id);
+            if (artist == null)
+            {
+                return NotFound();
+            }
+
             _context.Artists.Remove(artist);
-            _context.SaveChanges();
-            return _context.Artists.ToList();
+            await _context.SaveChangesAsync();
+
+            return Ok(artist);
+        }
+
+        private bool ArtistExists(int id)
+        {
+            return _context.Artists.Any(e => e.ArtistId == id);
         }
     }
 }
